first pass at discord bot stuff
